Update patch set 1
Patch Set 1: Code-Review-1

There is no part g and h on the agreed upon content of this document. This project, in sprint 1, agreed upon the structure and content of the requirements document. If you want to change that structure you should propose a change to the project and not simply add new sections to the document

Patch-set: 1
Reviewer: Gerrit User 1697 <1697@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1