Update patch set 2
Patch Set 2:

Can you clarify what you mean by section level?  

If you're talking about the section level, of the block of content then that is controlled by the RST markup following the section title.  I'm not a huge fan of this, but it's how it works.

You'll just want to look at a section header that is at the level you want this text to appear, and then use the characters it has to underline your section header.

For example:

Section Header
==============

Section Header 2
----------------

Patch-set: 2