Update patch set 1
Patch Set 1:

This fixes up the metadata for all the requirements that were updated or introduced in this release.  It also provides a conversion script to automatically perform this sync up in the future as well.

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0