Update patch set 1
Patch Set 1:

This update is ready for review.  Given where we are in Casablanca, I've opted to just update it in Dublin as it's primarily a move and a tweak from PNF to XNF requirements.  Do not attempt to cherry pick this back to Casablanca as it will likely fail.  If there's a desire to make this in Casablanca, then let me know

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0