Update patch set 1
Patch Set 1:

Rich - Can you tell me what log file on the build had this error?  It don't see the requirement ID duplicated our master, I don't get the error when I build locally off master, and the verify job did not have the issue.   It would seem to be local to the merge job, but I can't seem to find the workspace files or logs with the errors.

Patch-set: 1