Update patch set 2
Patch Set 2: Code-Review-1

(69 comments)

Still a lot of open issues. I'll setup a call for this week to discuss. Fair questions about what should be in VNF Requirements vs. the spec, but I'd like to handle that via a different ticket as there are already enough moving parts on this review as it is.

Please try to review comments in advance.  I'll be working through the review comments on the call.

Patch-set: 2
Label: Code-Review=-1