Update patch set 2
Patch Set 2:

Vijay - Are there any sections other than the schema changes that affect DCAE for Frankfurt?  Since there are no open issues against that, I think we can consider the schema final and move forward with it.  

The majority of the review comments open at this stage seem to be related more with the NF behavior or logistical issues such as where the schema should be stored - not listener behavior. I'd like to commit to getting everything finalized by EOW, but that may difficult for 100% of the items.  However, I think we can prioritize finalizing any items that affect DCAE.

Can you review the open comments and let us know which are most important to have closed and we'll start there?  We can discuss further on the review if needed.

Thanks,
Trevor

Patch-set: 2