Adding Schema VES Common Event for 7.1.1
NOTE: Corresponding spec updates will be published under a separate
review.  This is to allow DCAE to move forward with updates for
the ONAP Frankfurt release.  This file was previously reviewed
as part of VNFRQTS-802 and was broken out to allow further debate
on the other review feedback on that ticket

Issue-ID: VNFRQTS-814
Signed-off-by: Lovett, Trevor <trevor.lovett@att.com>
Change-Id: Ie38c80a12106372d326db6cdca38e005857d85f5