Update patch set 1
Patch Set 1:

All - Per Vijay's request to fast track the schema updates, I'm splitting this review into an independent update.  This is the same file that was reviewed on VNFRQTS-802.  I intend to merge this so DCAE can move forward with their development work.

Other items will be worked in the other review.

Patch-set: 1