Update patch set 4
Patch Set 4:

(31 comments)

Thanks to everyone for your continued support in getting this reviewed and resolved.  With the last commit I pushed up, I believe we have resolved or deferred every piece of feedback except for the message size issue.

I have an email chain going with the DCAE team to better understand the rationale for the revised 1MB limit.  If we determine there is no risk of failure or rejection of the message, then I will revise the language to state that 2MB is the maximum, but less than 1MB is recommended.   I will also note that the 1MB limit may be enforced in later version of the listener.

Based on the review this morning, I have created the following tickets to track the items we decided to defer.  These will be addressed in a future release of the spec:

* VNFRQTS-843
* VNFRQTS-844
* VNFRQTS-845
* VNFRQTS-846

Once the message size language is resolved, then I plan to merge this commit.  Any additional feedback should be submitted via a new VNFRQTS ticket.

Patch-set: 4