Update patch set 4
Patch Set 4:

I received feedback from both DCAE and DMAAP team that handle the ingestion and routing of the events, that the collector will not reject messages that are over 1MB, but DMAAP has stated in other contexts that messages over 1MB have lead to message loss and performance issues.

With this in mind, I will not change the maximum, **but** I will be putting language in that less than 1MB is strongly recommended and that a future version will enforce a 1MB limit.

Patch-set: 4
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0