Update patch set 1
Patch Set 1:

All - I am adding the VES 5.4.1 Event Listener Spec as we have determined that DCAE still supports it.  I have not made any changes to the content other that formating the line breaks and fixing some formatting that was causing some content not to be displayed.  This was copied from its prior location in VNFSDK from the Casablanca release.

I am not planning to entertain any content changes to the spec itself.

I've also updated R-570134 to denote that 5.4.1 is available, but 7.1.1 is preferred.

Please let me know if you have any issues or concerns.

Patch-set: 1