Update patch set 2
Patch Set 2:

All - 

Based on Vijay's comments, I think we should remove the unsupported features from the specification. We're in a rather awkward position with this as it's been published for quite some time, but I still think it's better to have accurate information in the spec than leave it or introduce a new version. So my plan is to remove/amend the content Vijay has pointed to even thought that's an odd action to take with an old baselined spec. Please let me know if anyone disagrees with this course of action.

So it's easier to track the changes I am making, we're going to merge this change in and I'll deliver the removals in the Frankfurt release via https://jira.onap.org/browse/VNFRQTS-861. I think this is better than removing sections in a ticket that was created to simply copy it forward.

Thanks,
Trevor

Patch-set: 2
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0