Relax camel casing requirements in VES
In 7.1, the VES spec added requirements that
made it mandatory for event names and
other fields to camel case the values
of those fields. This had the negative
side-effect of requiring vendor and
operator updates for existing VES
events and their associated processing.

To avoid this, the requirements have been
updated from must requirements to should

Issue-ID: VNFRQTS-869
Signed-off-by: Lovett, Trevor <trevor.lovett@att.com>
Change-Id: I8669e08f7c4cd2df44239ca2fa49421e011f2e1d