Update patch set 1
Patch Set 1:

(1 comment)

I'm OK pushing this into 7.1.1, but I do want to be clear on what that means to 3GPP.  Due to the timing of standard approval in 3GPP, we were pushed to provide a link to the specification prior to the end of the ONAP Frankfurt release.  I attempted to provide them a link to the document in Github using the Frankfurt branch, but both 3GPP and the ONAP TSC felt uncomfortable pointing at a branch that could technically change so they insisted on referring to a specific commit.

This means if we change this, the 3GGP spec will not see the change (at least without going through the whole notification process again and potentially impacting the relevant 3GPP documents as well).

This is primarily an informational change and does not materially change the specification.

I see 3 options:

1) We push it in, and leave 3GPP as-is
2) We defer this to 7.2
3) We push it in, and notify 3GPP of the change and try to bring them in sync

My personal vote is for #2, but I'm OK with #1.  I'm not in favor of #3, but if enough people are then I'll see what we can do.

Thanks,
Trevor

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>