Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Yep, Vijay is correct here.
> We`re assigining the schema file to the eventListener interface version in VES configuration.
> The schema file associated with VES Event Listener 7.2 spec needs to have a unique name.

It will have a unique name.  The intent of this commit was to establish a point to track changes.

If I change things as I'm copying things over to a new file, then it becomes very difficult to track what actually changed.  I don't mind changing the name of the file, but I think it would be easier to review if we track the content updates separately.

Patch-set: 1