Update patch set 1
Patch Set 1: Code-Review-1

This appears to be a backwards incompatible change.  It doesn't impact the schema, but the spec change could require changes to a client. I don't think this can go in a patch release.  

Given that Guilin and 7.2 is not yet officially published, should we just squeeze this in there and simply head off the issue?

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1